requirements.txt
requests
pymongo
elasticsearch
python-dotenv
flask
